% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows.R
\name{add_rows}
\alias{add_rows}
\title{Add Rows}
\usage{
add_rows(data_frame, first_col, last_col)
}
\arguments{
\item{data_frame}{dataframe to be added}

\item{first_col}{Name of the first column in the range}

\item{last_col}{Name of last column in the range}
}
\description{
This function adds all the rows in a dataframe within a specified range of columns and returns a data frame with one row and specified range of columns containing the sum
}
