% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{be}
\alias{be}
\title{Pathways Corpus - Belgium}
\format{A data frame including the following variables:
\describe{
  \item{qid}{identifier of written question}
  \item{qdate}{date of written question }
  \item{name}{name of MP}
  \item{mp_id}{identifier of MP}
  \item{party}{party of MP}
  \item{cio}{Citizen of Immigrant Origin status of MP}
  \item{text}{text of written question}
  \item{title}{title of written question}
  \item{area}{subject area of written question}
  \item{department}{adressed department of written question}
}}
\source{
\url{https://www.lachambre.be/}
}
\usage{
be
}
\description{
A dataset containing parliamentary written questions from the Chamber of Representatives, Belgium (2010-2014).
}
\keyword{datasets}
