% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{corpus_explorer}
\alias{corpus_explorer}
\title{launch the pathways shiny app}
\usage{
corpus_explorer(use_browser = TRUE)
}
\arguments{
\item{use_browser}{Choose whether you want to launch the shiny app in your browser.
Defaults to \code{TRUE}.}
}
\description{
\code{corpus_explorer} launches the shiny app to explore the Pathways corpus.
}
