% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_col_ndata.R
\name{del_cols_ndat}
\alias{del_cols_ndat}
\title{Delete all columns which have no data}
\usage{
del_cols_ndat(dataframe)
}
\arguments{
\item{dataframe}{Dataframe to be processed}
}
\description{
This function deletes all the columns in a dataframe which are completely empty
}
