% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_dup.R
\name{del_dup}
\alias{del_dup}
\title{Delete duplicates}
\usage{
del_dup(colname, data_frame)
}
\arguments{
\item{colname}{Name of the column containing duplicate names}

\item{data_frame}{Dataframe to be processed}
}
\value{
Dataframe with columns deleted
}
\description{
Delete rows which have duplicate cells in a particular column
}
