% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_rows_ndata.R
\name{del_rows_ndat}
\alias{del_rows_ndat}
\title{Delete rows which have no data between specified range of columns}
\usage{
del_rows_ndat(dataframe, startname, endname)
}
\arguments{
\item{dataframe}{Dataframe to be processed}

\item{startname}{Name of the first column in the range on columns}

\item{endname}{Name of the last colume in the range of columns}
}
\description{
This function deletes all rows which have no data within a range of columns specified using start column name and end column name
}
