% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{el}
\alias{el}
\title{Pathways Corpus - Greece}
\format{A data frame including the following variables:
\describe{
  \item{qid}{identifier of written question}
  \item{qdate}{date of written question }
  \item{name}{name of first MP}
  \item{mp_id}{identifier of MP}
  \item{party}{party of MP}
  \item{cio}{Citizen of Immigrant Origin status of MP}
  \item{text}{text of written question}
  \item{contributors}{names of other MPs}
  \item{area}{subject area of written question}
}}
\source{
\url{https://www.hellenicparliament.gr/}
}
\usage{
el
}
\description{
A dataset containing parliamentary written questions from the Hellenic Parliament, Greece (2015).
}
\keyword{datasets}
