% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{es}
\alias{es}
\title{Pathways Corpus - Spain}
\format{
A data frame including the following variables:
\describe{
  \item{qid}{identifier of written question}
  \item{qdate}{date of written question }
  \item{name}{name of MP}
  \item{mp_id}{identifier of MP}
  \item{party}{party of MP}
  \item{cio}{Citizen of Immigrant Origin status of MP}
  \item{text}{text of written question}
  \item{prestring}{prestring of written question}
  \item{contributors}{names of other MPs}
}
}
\source{
\url{http://www.congreso.es/}
}
\usage{
es
}
\description{
A dataset containing parliamentary written questions from the Congress of Deputies, Spain (2012-2015).
}
\keyword{datasets}
