% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_col.R
\name{fill_coll}
\alias{fill_coll}
\title{Fill cell between rows of a particular column which are NA with data in the row preceding empty space}
\usage{
fill_coll(x, colname)
}
\arguments{
\item{x}{Dataframe to be manupulated}

\item{colname}{String containing the name of the column to be manupulated}
}
\value{
Dataframe with the specified column filled
}
\description{
Fill cell between rows of a particular column which are NA with data in the row preceding empty space
}
