% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_text.R
\name{find_text}
\alias{find_text}
\title{Find text}
\usage{
find_text(text, data_frame)
}
\arguments{
\item{text}{Full or partial text to be found}

\item{data_frame}{Dataframe to be searched}
}
\value{
Matrix of row and colunm names of all cells containing the text
}
\description{
Find the all the the index of all the cells which contain full or partial text
}
