% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr}
\alias{fr}
\title{Pathways Corpus - France}
\format{A data frame including the following variables:
\describe{
  \item{qid}{identifier of written question}
  \item{qdate}{date of written question }
  \item{name}{name of MP}
  \item{mp_id}{identifier of MP}
  \item{party}{party of MP}
  \item{cio}{Citizen of Immigrant Origin status of MP}
  \item{text}{text of written question}
  \item{prestring}{prestring of written question}
  \item{area}{subject area of written question}
}}
\source{
\url{http://www.assemblee-nationale.fr/}
}
\usage{
fr
}
\description{
A dataset containing parliamentary written questions from the National Assembly, France (2007-2012).
}
\keyword{datasets}
