% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sen_meta.R
\name{gen_sen_meta}
\alias{gen_sen_meta}
\title{Generate scenario meta infotmation template}
\usage{
gen_sen_meta(ul_dfile_path, meta_template_path)
}
\arguments{
\item{ul_dfile_path}{Path to the data file for which the template is to be generated}

\item{meta_template_path}{Path to the default scenario meta template file in the csv format}
}
\description{
This function generates and writes a scenario meta information template for a particular data file with minimum information required to upload the template on the portal in the same directory as the data file
}
