% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{gene_sets_human}
\alias{gene_sets_human}
\title{Human Gene Sets}
\format{
\code{gene_sets_human} is a list with the following
  list elements:

\describe{

  \item{gene_info}{Data frame containing information on human
    genes, including gene symbols and Ensembl ids.}

  \item{gene_set_info}{Data frame contaiing information on gene
    sets, including gene set name, id and database of origin.}

  \item{gene_sets}{Gene sets encoded as a 61,676 x 37,856 sparse
    matrix, in which gene_sets[i,j] = 1 if gene j belongs to gene set
    i; otherwise, gene_sets[i,j] = 0.}}
}
\description{
37,856 gene sets compiled from NCBI BioSystems,
  Pathway Commons and MSigDB databases.
}
\keyword{data}
