% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{gene_sets_mouse}
\alias{gene_sets_mouse}
\title{Title Goes Here}
\format{
\code{gene_sets_mouse} is a list with the following
  list elements:

\describe{

  \item{item}{Describe item here.}

}
}
\description{
Add description here.
}
\keyword{data}
