% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it}
\alias{it}
\title{Pathways Corpus - Italy}
\format{A data frame including the following variables:
\describe{
  \item{qid}{identifier of written question}
  \item{qdate}{date of written question }
  \item{name}{name of first MP}
  \item{mp_id}{identifier of MP}
  \item{party}{party of MP}
  \item{cio}{Citizen of Immigrant Origin status of MP}
  \item{text}{text of written question}
  \item{contributors}{names of other MPs}
  \item{area}{subject area of written question}
}}
\source{
\url{http://dati.camera.it/}
}
\usage{
it
}
\description{
A dataset containing parliamentary written questions from the Chamber of Deputies, Italy (2008-2013).
}
\keyword{datasets}
