% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nl}
\alias{nl}
\title{Pathways Corpus - Netherlands}
\format{A data frame including the following variables:
\describe{
  \item{qid}{identifier of written question}
  \item{qdate}{date of written question }
  \item{name}{name of first MP}
  \item{mp_id}{identifier of MP}
  \item{party}{party of MP}
  \item{cio}{Citizen of Immigrant Origin status of MP}
  \item{text}{text of written question}
  \item{title}{title of written question}
  \item{contributors}{names of other MPs}
}}
\source{
\url{https://zoek.officielebekendmakingen.nl/}
}
\usage{
nl
}
\description{
A dataset containing parliamentary written questions from the House of Representatives, Netherlands (2010-2012).
}
\keyword{datasets}
