% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst.R
\name{pathways_analy}
\alias{pathways_analy}
\title{pathways_analy to pathways and analy}
\usage{
pathways_analy(
  data,
  group,
  org = "hsa",
  p_model = c("phyper", "fisher"),
  scale = T,
  p.adjust.methods = "holm",
  import_model = c("betweenness", "degree")
)
}
\arguments{
\item{data}{a list of data}

\item{group}{data group}

\item{org}{looking for https://www.genome.jp/kegg/catalog/org_list.html,using fread to keggAPI}

\item{scale}{scale of data ,It well to .get_data_analyst_plot}

\item{p.adjust.methods}{looking for  p.adjust}

\item{import_model}{import_model=c("betweenness","degree")}

\item{fisher.alternative}{alternative  of fisher.test}
}
\value{
metaProfiler data
}
\description{
pathways_analy to pathways and analy
}
\examples{

\dontrun{


x<-readxl::read_xlsx(system.file("data", "Col-top50.xlsx", package = "pathways"))
y<-readxl::read_xlsx(system.file("data", "metadata.xlsx", package = "pathways"))\%>\% as.data.frame()\%>\% na.omit
rownames(y)<-y$Sample\%>\%str_extract_all("[0-9]+[0-9]")\%>\% unlist \%>\% paste("cid:",., sep = "")
y[,-1]->y
library(biomaRt)
mart <- useMart("ensembl","mmusculus_gene_ensembl")
gene_name<-getBM(attributes=c("ensembl_gene_id","entrezgene_id"),
                 filters = "ensembl_gene_id",values = x$Gene, mart = mart)
inner_join(gene_name,x,by=c("ensembl_gene_id"="Gene"))\%>\% na.omit->getkegg
getkegg$ensembl_gene_id[!is.na(getkegg$entrezgene_id)]<-getkegg$entrezgene_id[!is.na(getkegg$entrezgene_id)]
getkegg[,-2]->genes
rownames(genes)<-genes$ensembl_gene_id
genes[,-1]->genes
group<-c(rep("A",4),rep("B",5))
list(a=y\%>\%t\%>\%as.data.frame(),b=genes\%>\%t\%>\%as.data.frame())->data
data\%>\% setNames(c("a","b"))->data
names<-match(data[[1]]\%>\% rownames,data[[2]]\%>\% rownames)
data[[2]]<-data[[2]][names,]
pathways_analy(data = data,group = group,org = "10090",scale = T)->n

# or
load(system.file("data", "Tes.Rdata",package = "pathways"))
}

}
