% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_facs_z.R
\docType{data}
\name{pbmc_facs_z}
\alias{pbmc_facs_z}
\title{Differential Expression Results from PBMC Data}
\format{
\code{pbmc_facs_z} is a matrix of z-scores, with rows
corresponding to (human) genes and columns corresponding to
differential expression analyses; differential expression was
measured separately for each cell type (B cells, T cells, natural
killer cells, CD14+ cells and CD34+ cells).
}
\description{
Results from differential expression analysis of
single-cell RNA-seq profiles generated from bead-enriched
subpopulations of peripheral blood mononuclear cell (PBMCs). The
data were downloaded from
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}.
}
\references{
Zheng et al (2017). Massively parallel digital transcriptional
profiling of single cells. \emph{Nature Communications} \bold{8},
14049. \url{https://doi.org/10.1038/ncomms14049}
}
\keyword{data}
