% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_plot.R
\name{plot_circos}
\alias{plot_circos}
\title{packagename: Package containing tools support project work.}
\usage{
plot_circos(result_data, FC)
}
\description{
This package provides a central repository for methods to facilitate
project work
}
\examples{

\dontrun{
library(tidyverse)
library(data.table)
library(clusterProfiler)
kegg_pathway1(data=list(a=c("cid:784","cid:962","cid:1004","cid:5862","cid:5886"),b=c("P0DTD3","Q9BYF1","Q9NRS4","Q9NYK1"),c=c("3.4.22.15","3.4.22.69","3.6.4.12","3.6.4.13"),
d=c("6921","6923","8453","8883","9039","9978","79699")))->da
save(da,file="pathways.RData")
 load("pathways.RData")
comp_dotplot(da@kegg_analyst$enrichKEGG$a)
da->lllll
lllll->da
 da@kegg_analyst$enrichKEGG$a@result\%>\% dplyr::filter(org=="KEGG") ->y
library(randomcoloR)
library(circlize)
library(ComplexHeatmap)
FC<-c(0.5,1.2,0.4,0.1,2)
names(FC)<-y$geneID\%>\%unlist\%>\% str_split("/")\%>\%unlist \%>\%unique
plot_circos(y,FC)
}
}
