% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_plot.R
\name{plot_funmap}
\alias{plot_funmap}
\title{pathways_analy to pathways and analy}
\usage{
plot_funmap(result_data, col = "auto")
}
\arguments{
\item{col}{data col like}

\item{data}{daat@kegg_analyst$enrichKEGG$d@result}
}
\value{
ggplot2
}
\description{
pathways_analy to pathways and analy
}
\examples{

\dontrun{ 
library(tidyverse)
library(data.table)
library(clusterProfiler)
kegg_pathway1(data=list(a=c("cid:784","cid:962","cid:1004","cid:5862","cid:5886"),b=c("P0DTD3","Q9BYF1","Q9NRS4","Q9NYK1"),c=c("3.4.22.15","3.4.22.69","3.6.4.12","3.6.4.13"),
d=c("6921","6923","8453","8883","9039","9978","79699")))->da
save(da,file="pathways.RData")
 load("pathways.RData")
comp_dotplot(da@kegg_analyst$enrichKEGG$a)
da->lllll
lllll->da
 da@kegg_analyst$enrichKEGG$a@result\%>\% dplyr::filter(org=="KEGG") ->y
 da@kegg_analyst$enrichKEGG$d@result->d23
plot_funmap(d23)
plot_funmap(d23,col=c("none"="red"))
# or
}
}
