% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{transition}
\alias{transition}
\title{Transition from model name to default names}
\usage{
transition(d_file, tr_file, notes, prod_ind_list)
}
\arguments{
\item{d_file}{Path to the raw data file}

\item{tr_file}{Path to the transition file}

\item{notes}{A binary value for the arguement note which generates a file to add conversion factors and notes}

\item{prod_ind_list}{Path to the file containing list of indicators on the portal currently in Emissions Scenarios>Indicators>ind_list_production.xlsx}
}
\value{
A data file with all the model indicator names replaced with default indicator names and all duplicates daletes. If notes is set to 1 a file to enter notes
}
\description{
Replaces model indicator names in the raw data file aready in the pathways template by default indicator names using the tranistion file matching both indicator names, deletes all duplicate values and generates a file to enter notes
}
