% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_with_add.R
\name{transition_with_add}
\alias{transition_with_add}
\title{Transition from model name to default names and adds 2 or more indicators if mentioned in the transition file}
\usage{
transition_with_add(d_file, tr_file, notes, prod_ind_list)
}
\arguments{
\item{d_file}{A string containing raw data file name without the extension}

\item{tr_file}{A string containing transition file name without the extension}

\item{notes}{A binary value for the arguement note which generates a file to add conversion factors and notes}

\item{prod_ind_list}{Path to the file containing list of indicators on the portal currently in Emissions Scenarios>Indicators>ind_list_production.xlsx}
}
\value{
A data file with all the model indicator names replaced with default indicator names and all duplicates daletes. If notes is set to 1 a file to enter notes
}
\description{
Replaces model indicator names in the raw data file aready in the pathways template by default indicator names using the tranistion file matching both indicator names, adds indicators which are mentioned in the transition file to be added, deletes all duplicate values and generates a file to enter notes
}
