% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gsea_plotly}
\alias{gsea_plotly}
\title{Add Title Here}
\usage{
gsea_plotly(
  gsea_res,
  gene_set_info,
  k,
  file,
  height = 675,
  width = 800,
  title = NULL,
  min_pval = 1,
  max_name_len = 44
)
}
\arguments{
\item{gsea_res}{Describe input argument "gsea_res" here.}

\item{gene_set_info}{Describe input argument "gene_set_info" here.}

\item{k}{Describe input argument "k" here.}

\item{file}{Describe input argument "file" here.}

\item{height}{Describe input argument "height" here.}

\item{width}{Describe input argument "width" here.}

\item{title}{Describe input argument "title" here.}

\item{min_pval}{Describe input argument "min_pval" here.}

\item{max_name_len}{Describe input argument "max_name_len" here.}
}
\value{
Describe return value here.
}
\description{
Create an interactive scatterplot using plotly to
explore the gene-set enrichment results for a given topic k. Input
argument "gene_set_info" is a data frame containing information
about the gene sets; "gsea_res" is an output from function
"perform_gsea"; and "label_gene_sets" is a vector of gene set ids
to be labeled in the plot.
}
\examples{
# See perform_gsea for examples.

}
\seealso{
\code{\link{perform_gsea}}
}
