% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{perform_gsea}
\alias{perform_gsea}
\title{Perform Gene Set Enrichment Analysis}
\usage{
perform_gsea(gene_sets, Z, verbose = TRUE, eps = 1e-32, nproc = 0, ...)
}
\arguments{
\item{gene_sets}{Gene set data encoded as an n x m binary matrix,
where n is the number of gene sets and m is the number of genes:
\code{gene_sets[i,j] = 1} if and only if gene i is included in gene
set j, and otherwise \code{gene_sets[i,j] = 0}. The rows and
columns should be named. Only the rows of \code{gene_sets} matching
the rows of \code{Z} will be used in the enrichment analysis.}

\item{Z}{Matrix of gene-level statistics such as z-scores, with
rows corresponding to genes. An enrichment analysis is performed
for each column of \code{Z}. The rows and columns should be named;
rows of \code{Z} matching the columns of \code{gene_sets} will
be used in the enrichment analysis.}

\item{verbose}{When \code{verbose = TRUE}, information about the
progress of the gene set enrichment analysis is printed to the
console.}

\item{eps}{The lower bound for calculating p-values; smaller values
of \code{eps} may give more accurate p-values at the possible cost
of slightly longer computation. Passed as the \code{eps} input to
\code{\link[fgsea]{fgsea}}.}

\item{nproc}{Number of workers used to run gene set enrichment
analysis. Passed as the \code{nproc} input to
\code{\link[fgsea]{fgsea}}.}

\item{\dots}{Additional arguments passed to
\code{\link[fgsea]{fgsea}}.}
}
\value{
The return value is a list containing four n x k matrices
of gene set enrichment analysis results, where n is the number of
gene sets and k is the number of columns in \code{Z}. The matrices
give the p-values (pval), enrichment scores (ES), normalized
enrichment scores (NES), and expected errors (log2err). See
\code{\link[fgsea]{fgseaMultilevel}} for more information about
these outputs.
}
\description{
A simple interface for performing gene set enrichment
  analysis on multiple collections of gene-level statistics.
}
\examples{
# Load the gene sets and the results of the differential
# expression analysis.
data(gene_sets_human)
data(pbmc_facs_z)

# Perform the gene set enrichment analysis using fgsea.
gsea_res <- perform_gsea(gene_sets_human$gene_sets,pbmc_facs_z)

# To improve the plot, add more info to the "database" column.
gene_set_info <-
  transform(gene_sets_human$gene_set_info,
            database = factor(paste(database,
              ifelse(database == "MSigDB",
                     as.character(category_code),
                     as.character(data_source)),sep = "-")))

# Create an interactive plot for exploring the results of the gene
# set enrichment analysis.
gsea_plotly(gsea_res,gene_sets_human$gene_set_info,"B",
            file = "gsea_b_cells.html",title = "B cells")

}
